 \name{Andes}
\alias{Andes}
\docType{data}
\title{Data collection of igneous rocks from Central Andes}
\description{
Compilation of more than 1500 major and trace element analysis of igneous rocks from Central Andes.
}
\usage{data(Andes)}
\format{
The format is a data frame with 1511 rows and 40 columns.
}
\details{This data set documents compositional variations of magmas since Jurassic time to Neogene period in the Andean orogeny.
}

\source{
The database is also available in GEOROC (Geochemistry of Rocks of the Oceans and Continents). For full details see \url{http://georoc.mpch-mainz.gwdg.de/georoc/} . 
}

\references{
Mamani, M., Worner, G., and Sempere, T. 2010. Geochemical variations in igneous rocks of the Central Andean orocline (13 S to 18 S): Tracing crustal thickening and magma generation through time and space. Bulletin of  Geological Society of America. v. 122, pp. 162--182.
}
\examples{
data(Andes)
##
}
\seealso{
\code{\link{TestTAS}}
}
\keyword{datasets}


